requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-mistralai==0.0.1
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
tiktoken==0.5.2
sentence-transformers==2.2.2
chromadb==0.4.18
faiss-cpu==1.7.4
pypdf==3.17.4
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.0
python-magic==0.4.27
python-magic-bin==0.4.14
textract==1.6.5
beautifulsoup4==4.12.2
streamlit==1.29.0
streamlit-extras==0.3.5
streamlit-option-menu==0.3.6
aiohttp==3.9.1
loguru==0.7.2
pydantic==2.5.0
pyyaml==6.0.1
tqdm==4.66.1
langdetect==1.0.9
polyglot==16.7.4
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3
readtime==3.0.0
dateutil==2.8.2
python-dateutil==2.8.2
filelock==3.13.1
pillow==10.1.0
openpyxl==3.1.2
xlrd==2.0.1
markdown==3.5.1
dotenv
pytesseract
opencv-python==4.8.1.78
Pillow==10.1.0
docx
langdetect
langchain
langchain-mistralai[pyright]
streamlit